Move buildSelectQueries to utils file
